Feat: 댓글 삭제/수정 기능 구현 + password로 권한 확인 코드 추가

+++ src/main/java/likelion14th/blog/service/ArticleService.java
@@ -2,20 +2,25 @@ package likelion14th.blog.service;
 
 import jakarta.persistence.EntityNotFoundException;
 import likelion14th.blog.domain.Article;
+import likelion14th.blog.domain.Comment;
+import likelion14th.blog.dto.request.CommentRequest;
 import likelion14th.blog.dto.response.ArticleDetailResponse;
 import likelion14th.blog.dto.response.ArticleSummaryResponse;
 import likelion14th.blog.repository.ArticleRepository;
+import likelion14th.blog.repository.CommentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional
 @RequiredArgsConstructor
 public class ArticleService {
     private final ArticleRepository articleRepository;
+    private final CommentRepository commentRepository;
 
     @Transactional
     public ArticleDetailResponse addArticle(String title, String content, String author, String password) {
@@ -34,9 +39,12 @@ public class ArticleService {
     }
 
     @Transactional
-    public ArticleDetailResponse updateArticle(Long id, String title, String content) {
+    public ArticleDetailResponse updateArticle(Long id, String title, String content, String password) {
         Article article = articleRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("해당 ID의 게시글을 찾을 수 없습니다."));
 
+        if(!article.getPassword().equals(password)) {
+            throw new RuntimeException("해당 게시글에 대한 수정 권한이 없습니다.");
+        }
         article.update(title, content);
 
         articleRepository.save(article); // 안해줘도 된다. post는 영속성 컨텍스트에 의해서 한번도 사용한 적이 없기 때문에 해줘야 하지만 patch는 자동으로 해준다.
@@ -44,7 +52,16 @@ public class ArticleService {
     }
 
     @Transactional
-    public Void deleteArticle(Long id) {
+    public Void deleteArticle(Long id, String password) {
+        Article article = articleRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("해당 게시글을 조회할 수 없습니다."));
+
+        if (!article.getPassword().equals(password)) {
+            throw new RuntimeException("해당 게시글에 대한 삭제 권한이 없습니다.");
+        }
+
+        List<Comment> comments = commentRepository.findByArticle(article);
+        comments.forEach(comment -> commentRepository.delete(comment));
+
         articleRepository.deleteById(id);
         return null;
     }
